requirements.txt
bokeh==2.4.3
dask>=2023.3.2
dask-jobqueue>=0.8.1
distributed>=2023.3.2
nibabel>=4.0.2
nilearn>=0.9.2
numpy>=1.23.4
pandas>=1.5.1
PyYAML>=6.0
scipy>=1.9.3
